% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyTranslations.R
\name{copyTranslations}
\alias{copyTranslations}
\title{Copy and edit translation files}
\usage{
copyTranslations(
  components,
  to.language,
  from.project,
  from.language,
  conflict = "ignore",
  verbose = FALSE,
  download = TRUE,
  download.method = "api",
  upload = TRUE,
  editFunction = editTranslationFile,
  ...
)
}
\arguments{
\item{components}{A list of component names (slugs).}

\item{to.language}{The language to copy to.}

\item{from.project}{The project to copy from.}

\item{from.language}{The language to copy from.}

\item{conflict}{Optional. How to handle conflicts on the server. One of "ignore", "replace-translated" or "replace-approved".}

\item{verbose}{Optional. Whether to print a detailed log to the console or not.}

\item{download}{Optional. Whether to download translation files (uses files on disk if FALSE).}

\item{download.method}{Optional. Which method to use to download the translation files. See \code{\link{getFile}}.}

\item{upload}{Optional. Whether to upload translation files after download (and editing).}

\item{editFunction}{Optional. The function to edit the translation files. Defaults to \code{\link{editTranslationFile}}. Needs to have at least the parameters "slug" and "from.language", save processed files as "slug-i.csv" (with i in 1:n) and return n the number of created files. During upload, all of processed files will be uploaded.}

\item{...}{Parameters to be passed to the editFunction. See \code{\link{editTranslationFile}}}
}
\value{
Count of accepted new translations in the destination project.
}
\description{
Download translation files for a Weblate project, edit them and then reupload them to the same or a different Weblate project.
}
\examples{
\dontrun{
updateCount <- copyTranslations(components = "address",
                                to.language = "de_CH",
                                from.project = "gpde",
                                from.language = "de",
                                filter = "Deutschland",
                                replace = as.data.frame(cbind(pattern = c("ß"),
                                                              replace = c("ss"))))
}
}
