% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{deleteTranslations}
\alias{deleteTranslations}
\title{Delete translations for a component on the server}
\usage{
deleteTranslations(slug, project, language, verbose = FALSE)
}
\arguments{
\item{slug}{The component to delete from}

\item{project}{The project to delete form}

\item{language}{The language to delete from}

\item{verbose}{Optional. Whether to print a detailed log to the console or not.}
}
\value{
The response from the server, as returned by httr::DELETE().
}
\description{
Deletes the translations of the specified component on the server.
}
