% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{getComponents}
\alias{getComponents}
\title{Get all components within a project.}
\usage{
getComponents(project)
}
\arguments{
\item{project}{The project you want to get the list of components from.}
}
\value{
A data frame with names and slugs for all components in the project.
}
\description{
Get all components within a project.
}
