% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{postFile}
\alias{postFile}
\title{Upload a translation file}
\usage{
postFile(
  slug,
  to.language,
  from.directory,
  conflict = "ignore",
  filename,
  verbose = FALSE
)
}
\arguments{
\item{slug}{The slug of the component to upload to.}

\item{to.language}{The language to upload to.}

\item{from.directory}{The directory the file is saved in.}

\item{conflict}{Optional. How to handle conflicts on the server. One of "ignore", "replace-translated" or "replace-approved".}

\item{filename}{Optional, defaults to \link{slug}. Name of the file, without ".csv" ending.}

\item{verbose}{Optional. Whether to print a detailed log to the console or not.}
}
\value{
The response from the server, as returned by httr::POST().
}
\description{
Upload the translation file stored at "~/\link{from.language}/\link{slug}.csv" to the
specified language in the project defined in setup().
}
