% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Initialize some basic environment variables needed for the API calls.}
\usage{
setup(api.url, token, to.project)
}
\arguments{
\item{api.url}{The URL of your API entry point. Usually ""http://example.com/api/".}

\item{token}{Your authorization token. See https://docs.weblate.org/en/latest/api.html#authentication-and-generic-parameters}

\item{to.project}{The project you want to upload the translation files to. CAREFUL: make sure to set this to the appropriate project.}
}
\value{

}
\description{
Initialize some basic environment variables needed for the API calls.
}
