% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyTranslations.R
\name{copyTranslations}
\alias{copyTranslations}
\title{Copy and edit translation files}
\usage{
copyTranslations(
  components,
  to.language,
  from.project,
  from.language,
  filter,
  replace,
  conflict = "ignore"
)
}
\arguments{
\item{components}{A list of component names (slugs).}

\item{to.language}{The language to copy to.}

\item{from.project}{The project to copy from.}

\item{from.language}{The language to copy from.}

\item{filter}{List of strings. Translations containing any of these strings will not be reuploaded.}

\item{replace}{Data frame of search/replace string pairs.}

\item{conflict}{How to handle conflicts on the server. One of "ignore", "replace-translated" or "replace-approved".}
}
\value{
Count of accepted new translations in the destination project.
}
\description{
Download translation files for a Weblate project, edit them (search/replace &
filter unwanted terms) and then reupload them to the same or a different
Weblate project.
}
\examples{
\dontrun{
updateCount <- copyTranslations(components = "address",
                                to.language = "de_CH",
                                from.project = "GPDE",
                                from.language = "de",
                                filter = "Deutschland",
                                replace = as.data.frame(cbind(pattern = c("ß"),
                                                              replace = c("ss"))))
}
}
