% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyTranslations.R
\name{copyTranslations}
\alias{copyTranslations}
\title{Copy and edit translation files}
\usage{
copyTranslations(
  components,
  to.language,
  from.project,
  from.language,
  filter,
  replace,
  conflict = "ignore",
  size.limit = 8e+05,
  verbose = FALSE,
  download = TRUE,
  upload = TRUE,
  editFunction = editTranslationFile
)
}
\arguments{
\item{components}{A list of component names (slugs).}

\item{to.language}{The language to copy to.}

\item{from.project}{The project to copy from.}

\item{from.language}{The language to copy from.}

\item{filter}{Optional. List of strings. Translations containing any of these strings will not be reuploaded.}

\item{replace}{Optional. Data frame of search/replace string pairs. Needs to contain a column "pattern" and a column "replace".}

\item{conflict}{Optional. How to handle conflicts on the server. One of "ignore", "replace-translated" or "replace-approved".}

\item{size.limit}{The servers file size limit for uploads. Defaults to 800000. Decrease if getting HTTP 413 error on writing.}

\item{verbose}{Optional. Whether to print a detailed log to the console or not.}

\item{download}{Whether to download translation files (uses files on disk if FALSE).}

\item{upload}{Whether to upload translation files after download (and editing).}

\item{editFunction}{Option to pass a custom function for how to edit the translation files. Needs to have parameters (slug, from.language, filter, replace, size.limit), save processed files as "slug-x.csv" and return the number of files.}
}
\value{
Count of accepted new translations in the destination project.
}
\description{
Download translation files for a Weblate project, edit them (search/replace &
filter unwanted terms) and then reupload them to the same or a different
Weblate project.
}
\examples{
\dontrun{
updateCount <- copyTranslations(components = "address",
                                to.language = "de_CH",
                                from.project = "gpde",
                                from.language = "de",
                                filter = "Deutschland",
                                replace = as.data.frame(cbind(pattern = c("ß"),
                                                              replace = c("ss"))))
}
}
