% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editTranslations.R
\name{editTranslationFile}
\alias{editTranslationFile}
\title{Edit translation files}
\usage{
editTranslationFile(slug, from.language, filter, replace, size.limit = 8e+05)
}
\arguments{
\item{slug}{The slug of the component the translation file belongs to.}

\item{from.language}{The source language of the translation file.}

\item{filter}{Optional. List of strings. Translations containing any of these strings will be removed from the translation file.}

\item{replace}{Optional. Data frame of search/replace string pairs. Needs to contain a column "pattern" and a column "replace". Translation strings will be edited according to these patterns.}

\item{size.limit}{The servers file size limit for uploads. Defaults to 800000. Decrease if getting HTTP 413 error on writing. Translation files bigger than this will be split into multiple smaller files.}
}
\value{
The number of saved processed files.
}
\description{
Takes one translation file from disk, removes unwanted translations, edits translations according to provided patterns, splits the file into multiple if too big, saves to disk, and returns the number of created files.
}
