% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{getFile}
\alias{getFile}
\title{Download a translation file.}
\usage{
getFile(slug, from.project, from.language, verbose = FALSE)
}
\arguments{
\item{slug}{The slug of the component to download the translation file from.}

\item{from.project}{The project to download the file from.}

\item{from.language}{The language to download the file from.}

\item{verbose}{Optional. Whether to print a detailed log to the console or not.}
}
\value{
The response from the server, as returned by httr::GET().
}
\description{
Download a translation file from the location specified by a combination of
Project, Component, and Language.
The file is written to disk at "~/\link{from.language}/\link{slug}.csv". Existing files
are overwritten.
}
