% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Initialize some basic environment variables needed for the API calls.}
\usage{
setup(base.url, token, to.project, debug = FALSE)
}
\arguments{
\item{base.url}{The URL of your Weblate instance. This is the same URL that you would use to access the Weblate web interface in your browser. "https://example.com" (no trailing "/").}

\item{token}{Your authorization token. See https://docs.weblate.org/en/latest/api.html#authentication-and-generic-parameters}

\item{to.project}{The project you want to upload the translation files to. CAREFUL: make sure to set this to the appropriate project. It is set here instead of in copyTranslations() or postFile() to avoid mistakenly writing to the wrong project.}

\item{debug}{Controls httr verbose logging (httr::verbose()). Defaults to FALSE}
}
\value{

}
\description{
Initialize some basic environment variables needed for the API calls.
}
